@@ -18,26 +18,27 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_map: ^0.12.0
+  flutter_map: 
+  latlong2: ^0.8.1
   dart_jts:
-    git:
-      url: https://github.com/bettdouglas/dart_jts.git
+    # git:
+    #   url: https://github.com/bettdouglas/dart_jts.git
   flutter_map_tappable_polyline:
-  flutter_riverpod: ^0.14.0+3
+  flutter_riverpod: ^2.1.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  backdrop: ^0.8.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  test:
   peanut: 
 
 # For information on the generic Dart part of this file, see the
